--- a/lib/src/importer/host.dart
+++ b/lib/src/importer/host.dart
@@ -2,8 +2,7 @@
 // MIT-style license that can be found in the LICENSE file or at
 // https://opensource.org/licenses/MIT.
 
-// ignore: deprecated_member_use
-import 'dart:cli';
+import 'dart:async';
 
 import 'package:sass_api/sass_api.dart' as sass;
 
@@ -23,9 +22,8 @@ class HostImporter extends ImporterBase {
   HostImporter(Dispatcher dispatcher, this._compilationId, this._importerId)
       : super(dispatcher);
 
-  Uri? canonicalize(Uri url) {
-    // ignore: deprecated_member_use
-    return waitFor(() async {
+  FutureOr<Uri?> canonicalize(Uri url) {
+    return (() async {
       var response = await dispatcher
           .sendCanonicalizeRequest(OutboundMessage_CanonicalizeRequest()
             ..compilationId = _compilationId
@@ -46,9 +44,8 @@ class HostImporter extends ImporterBase {
     }());
   }
 
-  sass.ImporterResult? load(Uri url) {
-    // ignore: deprecated_member_use
-    return waitFor(() async {
+  FutureOr<sass.ImporterResult?> load(Uri url) {
+    return (() async {
       var response =
           await dispatcher.sendImportRequest(OutboundMessage_ImportRequest()
             ..compilationId = _compilationId