@ntkme 2022-12-25
This actually turned out to regress if a function is passed but never used on a large user input due to most part of the compiler is running in event loop because dart simply doesn't optimize synchronously returned `Value` for `FutureOr<Value>` and still put it on event loop: https://github.com/dart-lang/language/issues/2033